# Make big Int stringification 1.58x as fast