@zoffixznet 2018-10-16
> basically by dividing by a larger number at once,

Yes, instead of doing seventeen `mp_div_d` divisions by 10, we now do a single `mp_div_d` division by 100_000_000_000_000_000, which is the largest 10ⁿ number that fits into `mp_digit` with our current configuration of libtommath (that's what `DIGIT_BIT = 60` conditionals are for), so then we can repeat these divisions as long as we have more than one `mp_digit` in our number.

> Do we have specific tests for the larger numbers?

Very likely, as many tests use `is` which stringifies. I also added an explicit one, to be sure: https://github.com/perl6/roast/commit/5ef9b0c504